// src/types/core.ts
export interface TimeRange {
  startMs: number;
  endMs: number;
}